fix some bugs

+++ Tests/LeatherbackTests/Assets/Leatherback/MetaDataBuffer/MetaDataBuffer_Deciphering.cs
@@ -47,9 +47,9 @@ namespace Leatherback {
 
             return val;
         }
-        public System.UInt32 GetUInt16 () { CheckBuffers();
+        public System.UInt16 GetUInt16 () { CheckBuffers();
 
-            System.UInt32 val = BitConverter.ToUInt16(byteBuffer, index); index += 2;
+            System.UInt16 val = BitConverter.ToUInt16(byteBuffer, index); index += 2;
 
             return val;
         }
